BlogCreate

+++ app/Http/Controllers/ArticleController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Article;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 class ArticleController extends Controller
 {
+    // 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
+    public function __construct(){
+        $this->middleware('auth')->except(['index','detail']);
+    }
     public function index(){
       $data= Article::latest()->Paginate(5);
       return view('index',['articles'=>$data]);
@@ -14,6 +20,7 @@ class ArticleController extends Controller
     }
     public function detail($id){
         $data=Article::find($id);
+        // dd($data->toArray());
         return view('detail',[
             'article'   =>  $data
         ]);
@@ -21,15 +28,29 @@ class ArticleController extends Controller
     }
     public function delete($id){
         $article=Article::find($id);
-        $article->delete();
+        // dd(auth()->user()->id);
+        // dd($article->toArray());
+        if($article->user_id===auth()->user()->id){
+            $article->delete();
+            return redirect()->route('index');
+        }else{
+            return back()->with('errors','Unauthorized');
+        }
+    
 
-       return redirect('/articles')->with('info', 'Article deleted');
-    }
+   
+   }
+
+    
+
+
+
     public function add(){
       $data=[
         [
             'id'    =>  1 , "name"  =>  "New ",
             'id'    =>  2 , "name"  =>  "Tech ",
+           
         ]
         ];
         return view('/add',[
@@ -42,6 +63,9 @@ class ArticleController extends Controller
             'title'         =>  'required',
             'body'          =>  'required',
             'category_id'   =>  'required',
+         
+          
+          
         ]);
         if($validator->fails()){
             return back()->withErrors($validator);
@@ -50,7 +74,16 @@ class ArticleController extends Controller
         $article->title= request()->title;
         $article->body=request()->body;
         $article->category_id=request()->category_id;
+        $article->user_id=Auth::id();
+    
+
         $article->save();
         return redirect('/articles');
+        
     }
+    
+
+     
+    
+    
 }
